switched INT to FLOAT price

+++ models/Item.py
@@ -5,7 +5,7 @@ class Item(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     item_name = db.Column(db.String(64))
-    item_price = db.Column(db.Integer)
+    item_price = db.Column(db.Float)
     order_item = db.relationship("Order_Item", backref="item", cascade="all, delete-orphan")
 
     def __repr__(self):
